refactor: DetailClub 본문 수정
