+++ src/pages/clubDetailPage/components/contentContainer/Content.styled.js
@@ -1,0 +1,66 @@
+import styled from "styled-components";
+
+export const PageContainer = styled.div`
+  min-height: 466px;
+  width: 590px;
+  margin-bottom: 65px;
+`;
+export const Intro = styled.div`
+  margin-bottom: 50px;
+  .title {
+    font-size: 20px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+`;
+export const IntroText = styled.div`
+  white-space: pre-wrap; /* 엔터키 반영 */
+  word-wrap: break-word; /* 긴 단어 줄바꿈 */
+  line-height: 1.5; /* 줄 간격 */
+  font-size: 15px;
+`;
+
+export const Goal = styled.div`
+  margin-bottom: 50px;
+  .title {
+    font-size: 20px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+`;
+export const GoalText = styled.div`
+  white-space: pre-wrap; /* 엔터키 반영 */
+  word-wrap: break-word; /* 긴 단어 줄바꿈 */
+  line-height: 1.5; /* 줄 간격 */
+  font-size: 15px;
+`;
+
+export const Plan = styled.div`
+  margin-bottom: 50px;
+  .title {
+    font-size: 20px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+`;
+export const PlanText = styled.div`
+  white-space: pre-wrap; /* 엔터키 반영 */
+  word-wrap: break-word; /* 긴 단어 줄바꿈 */
+  line-height: 1.5; /* 줄 간격 */
+  font-size: 15px;
+`;
+
+export const Progress = styled.div`
+  margin-bottom: 50px;
+  .title {
+    font-size: 20px;
+    font-weight: bold;
+    margin-bottom: 20px;
+  }
+`;
+export const ProgressText = styled.div`
+  white-space: pre-wrap; /* 엔터키 반영 */
+  word-wrap: break-word; /* 긴 단어 줄바꿈 */
+  line-height: 1.5; /* 줄 간격 */
+  font-size: 15px;
+`;
