+++ src/pages/clubDetailPage/components/floatingContainer/FloatingContainer.styled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const PageContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 33px;
+  width: 450px;
+`;
