+++ src/pages/clubDetailPage/components/floatingContainer/Profile.styled.js
@@ -1,0 +1,58 @@
+import styled from "styled-components";
+
+export const Seller = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  border: 1px solid #d9d9d9;
+  border-radius: 10px;
+  padding-top: 40px;
+  padding-bottom: 30px;
+  padding-left: 45px;
+  padding-right: 45px;
+`;
+export const ProfileWrap = styled.div`
+  display: flex;
+  align-items: center;
+  margin-bottom: 30px;
+  gap: 53px;
+  img {
+    width: 55px;
+    height: 55px;
+  }
+`;
+export const Pwrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+  .name {
+    font-weight: bold;
+    font-size: 20px;
+  }
+`;
+export const Email = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 53px;
+  font-size: 20px;
+  .img {
+    width: 55px;
+    height: 55px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+`;
+export const Heart = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 53px;
+  font-size: 20px;
+  .img {
+    width: 55px;
+    height: 55px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+`;
