+++ src/router.jsx
@@ -4,7 +4,7 @@ import MainPage from "./pages/mainPage/mainPage";
 import ClubPage from "./pages/clubPage/ClubPage";
 import PaperPage from "./pages/paperPage/PaperPage";
 import MyPage from "./pages/myPage/MyPage";
-import DetailClub from "./pages/clubPage/DetailClub";
+import DetailClub from "./pages/clubDetailPage/DetailClub";
 
 export default createBrowserRouter([
   {
